+++ examples/data-sources/langsmith_organizations/data-source.tf
@@ -1,0 +1,5 @@
+data "langsmith_organizations" "all" {}
+
+output "organization_ids" {
+  value = [for o in data.langsmith_organizations.all.organizations : o.id]
+}